bot/requirements.txt
python-telegram-bot==21.6
motor==3.6.0
pydantic==2.8.2
python-dotenv==1.0.1
